+++ twine-components/src/interpolation.rs
@@ -1,143 +1,3 @@
-use ndarray::Array1;
-use ninterp::{
-    error::{InterpolateError, ValidateError},
-    prelude::{Interp1DOwned, Interpolator},
-    strategy,
-};
-use thiserror::Error;
-use twine_core::Component;
-
-#[derive(Error, Debug)]
-pub enum InterpError {
-    #[error(transparent)]
-    Validation(#[from] ValidateError),
-    #[error(transparent)]
-    Interpolation(#[from] InterpolateError),
-}
-
-/// Extrapolation strategy
-///
-/// Controls what happens if supplied interpolant point is outside the bounds of
-/// the interpolation grid.
-#[derive(Clone, Copy, Debug, PartialEq, Default)]
-pub enum Extrapolate<T> {
-    /// Evaluate beyond the limits of the interpolation grid.
-    Enable,
-    /// If point is beyond grid limits, return this value instead.
-    Fill(T),
-    /// Restrict interpolant point to the limits of the interpolation grid,
-    /// using [`num_traits::clamp`].
-    Clamp,
-    /// Wrap around to other end of periodic data. Does NOT check that first and
-    /// last values are equal.
-    Wrap,
-    /// Return an error when interpolant point is beyond the limits of the
-    /// interpolation grid.
-    #[default]
-    Error,
-}
-
-impl<T> From<Extrapolate<T>> for ninterp::interpolator::Extrapolate<T> {
-    fn from(value: Extrapolate<T>) -> Self {
-        match value {
-            Extrapolate::Enable => ninterp::interpolator::Extrapolate::Enable,
-            Extrapolate::Fill(val) => ninterp::interpolator::Extrapolate::Fill(val),
-            Extrapolate::Clamp => ninterp::interpolator::Extrapolate::Clamp,
-            Extrapolate::Wrap => ninterp::interpolator::Extrapolate::Wrap,
-            Extrapolate::Error => ninterp::interpolator::Extrapolate::Error,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-pub enum Strategy {
-    Linear,
-    Nearest,
-    LeftNearest,
-    RightNearest,
-}
-
-pub struct Interp1D(Interp1DOwned<f64, strategy::enums::Strategy1DEnum>);
-
-impl Interp1D {
-    #[allow(clippy::missing_errors_doc)]
-    pub fn new<T: Into<Array1<f64>>>(
-        x: T,
-        f_x: T,
-        strategy: &Strategy,
-        extrapolate: Extrapolate<f64>,
-    ) -> Result<Self, InterpError> {
-        match strategy {
-            Strategy::Linear => Ok(Self(Interp1DOwned::new(
-                x.into(),
-                f_x.into(),
-                ninterp::strategy::Linear.into(),
-                extrapolate.into(),
-            )?)),
-            Strategy::Nearest => Ok(Self(Interp1DOwned::new(
-                x.into(),
-                f_x.into(),
-                ninterp::strategy::Nearest.into(),
-                extrapolate.into(),
-            )?)),
-            Strategy::LeftNearest => Ok(Self(Interp1DOwned::new(
-                x.into(),
-                f_x.into(),
-                ninterp::strategy::LeftNearest.into(),
-                extrapolate.into(),
-            )?)),
-            Strategy::RightNearest => Ok(Self(Interp1DOwned::new(
-                x.into(),
-                f_x.into(),
-                ninterp::strategy::RightNearest.into(),
-                extrapolate.into(),
-            )?)),
-        }
-    }
-}
-
-impl Component for Interp1D {
-    type Input = f64;
-    type Output = f64;
-    type Error = InterpError;
-
-    fn call(&self, input: Self::Input) -> Result<Self::Output, Self::Error> {
-        self.0.interpolate(&[input]).map_err(Into::into)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use twine_core::Component;
-
-    use super::*;
-
-    #[test]
-    fn strategy_interp_matches_expected_value() {
-        let test_cases = [
-            (Strategy::Linear, 0.56),
-            (Strategy::Nearest, 0.4),
-            (Strategy::LeftNearest, 0.4),
-            (Strategy::RightNearest, 0.8),
-        ];
-
-        for (strategy, expected) in test_cases {
-            let interp = Interp1D::new(
-                vec![0., 1., 2.],
-                vec![0.0, 0.4, 0.8],
-                &strategy,
-                Extrapolate::Error,
-            )
-            .unwrap();
-
-            let actual = interp.call(1.4).unwrap();
-            assert!(
-                approx::relative_eq!(actual, expected),
-                "strategy {:?} produced wrong result: got {}, expected {}",
-                strategy,
-                actual,
-                expected
-            );
-        }
-    }
-}
+pub mod error;
+pub mod extrapolate;
+pub mod one;
