+++ twine-components/src/interpolation/error.rs
@@ -1,0 +1,10 @@
+use ninterp::error::{InterpolateError, ValidateError};
+use thiserror::Error;
+
+#[derive(Error, Debug)]
+pub enum InterpError {
+    #[error(transparent)]
+    Validation(#[from] ValidateError),
+    #[error(transparent)]
+    Interpolation(#[from] InterpolateError),
+}
